requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
email-validator==2.0.0
openai>=1.50.0
requests==2.31.0
gunicorn==21.2.0
weasyprint==60.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0 